requirements.txt
streamlit
python-docx
rispy
PyMuPDF
